combined reducers for search, update, delete user
