+++ src/reducers/index.js
@@ -1,7 +1,5 @@
-import searchUserReducer from "../reducers/searchUserReducer";
+import userReducer from "../reducers/UserReducer";
 import reducerNewSystemUser from "../reducers/reducerNewSystemUser";
-import updateSystemUser from "../reducers/reducerUpdateSystemUser";
-import deleteSystemUser from "../reducers/deleteSystemUser";
 import { combineReducers } from "redux";
 import loginReducer from "../reducers/loginReducer";
 import reducerProductStatsVue from "../reducers/reducerProductStatsVue";
@@ -21,10 +19,8 @@ import itemTypeReducer from "./itemTypeReducer.js";
 // import getSentiment from '../reducers/sentimentReducer';
 
 export default combineReducers({
-  searchUserReducer: searchUserReducer,
+  userReducer: userReducer,
   addSystemUserReducer: reducerNewSystemUser,
-  updateSystemUser: updateSystemUser,
-  deleteSystemUser: deleteSystemUser,
   login: loginReducer,
   prodStatData: reducerProductStatsVue,
   farmerCardData: reducerFarmerCardPreviews,
